Add examples of generator / async interop.
This behavior is technically deprecated as of Python 3.9, but it
does still work and is helpful to understand.

My opinion is that pyre is probably not using reasonable types
in its representation of asyncio coroutines: we're currently using
a Coroutine[None, None, ReturnType], but really an awaitable is
a distinct type Awaitable[ReturnType]. It's not a true coroutine,
because yields are illegal and the result is not iterable.

The @asyncio.coroutine decorator should be treated as a function
that maps functions returning Coroutine[None, None, ReturnType] to
functions returning Awaitable[ReturnType], and also creates special
magic inside of the function so that `yield from <awaitable>`, which
normally produces a TypeError, becomes legal.