Add some tests of yield from behavior.
I usually take it for granted that bugs in CPython are vanishingly
rare, but I think I may have found one.

The CPython unit tests never verify behavior of a `yield from`
expression where there are yield statements in the outer generator
(the generator from which `yield from` is used).

My tests seem to indicate that the existence of outer yields...
I haven't yet verified whether it before vs after the yield from
matters... break the yield from behavior so that sends are not
proxied and neither are yields.

This does not seem to me like reasonable behavior for any programming
language, it means the evaluation of the expression is context
dependent in a weird and uncomposeable way - my impression from PEP 380
was that composeability is the whole point.

Regardless, it's nice to have this documented - if it's intended then
that's one hell of a sharp edge in the language.