Update examples - I had misunderstood send and next behavior
I was confused before. Now my tests make sense.

The fact that you always have to start up a coroutine with `next(coroutine)`,
combined with the fact that some tutorials only show one-way
communication, had led me to mistakenly conclude that `send` doesn't
actually return a value or advance control in the calling function. So
I thought you had to alternate between `send` and `next` calls.

That doesn't really make any sense, and it was a gross misread of PEP
342. It doesn't make sense because:
- if you send data it has to go somewhere, so the obvious thing is to
  reenter the generator at that point
- it would make no sense to have separate send and next functions that
  have to alternate, when combining them would produce a less bug-prone
  API.

And indeed, you have to call next() once on any coroutine (or
equivalently send(None)) in order to advance control flow, but after
that the `send` calls do advance control flow and do return the next
value to the caller.

This does make the typing rules a little weird - without some kind of
linear type (which Python doesn't have) you have to allow a `next` or
`send(None)` call on an arbitrary generator, even if the input type
is non-null. So it's basically impossible to have completely sound
typing of general coroutines. But the runtime behavior makes perfect
sense.